+++ tests/entrypoints/openai/chat_completion/test_harmony.py
@@ -1,220 +1,23 @@
 # SPDX-License-Identifier: Apache-2.0
 # SPDX-FileCopyrightText: Copyright contributors to the vLLM project
 
-import pytest
 from openai_harmony import Message, Role
 
 from tests.entrypoints.openai.utils import verify_harmony_messages
-from vllm.entrypoints.openai.parser.harmony_utils import (
+from vllm.entrypoints.openai.chat_completion.harmony import (
     auto_drop_analysis_messages,
-    get_encoding,
-    get_system_message,
-    has_custom_tools,
-    parse_chat_input_to_harmony_message,
+    chat_input_to_harmony,
     parse_chat_output,
 )
-from vllm.entrypoints.openai.responses.harmony import (
-    response_previous_input_to_harmony,
+from vllm.entrypoints.openai.harmony import (
+    get_encoding,
 )
-
-
-class TestCommonParseInputToHarmonyMessage:
-    """
-    Tests for scenarios that are common to both Chat Completion
-    parse_chat_input_to_harmony_message and Responses API
-    response_previous_input_to_harmony functions.
-    """
-
-    @pytest.fixture(
-        params=[parse_chat_input_to_harmony_message, response_previous_input_to_harmony]
-    )
-    def parse_function(self, request):
-        return request.param
-
-    def test_assistant_message_with_tool_calls(self, parse_function):
-        """Test parsing assistant message with tool calls."""
-        chat_msg = {
-            "role": "assistant",
-            "tool_calls": [
-                {
-                    "function": {
-                        "name": "get_weather",
-                        "arguments": '{"location": "San Francisco"}',
-                    }
-                },
-                {
-                    "function": {
-                        "name": "search_web",
-                        "arguments": '{"query": "latest news"}',
-                    }
-                },
-            ],
-        }
-
-        messages = parse_function(chat_msg)
-
-        assert len(messages) == 2
-
-        # First tool call
-        assert messages[0].author.role == Role.ASSISTANT
-        assert messages[0].content[0].text == '{"location": "San Francisco"}'
-        assert messages[0].channel == "commentary"
-        assert messages[0].recipient == "functions.get_weather"
-        assert messages[0].content_type == "json"
-
-        # Second tool call
-        assert messages[1].author.role == Role.ASSISTANT
-        assert messages[1].content[0].text == '{"query": "latest news"}'
-        assert messages[1].channel == "commentary"
-        assert messages[1].recipient == "functions.search_web"
-        assert messages[1].content_type == "json"
-
-    def test_assistant_message_with_empty_tool_call_arguments(self, parse_function):
-        """Test parsing assistant message with tool call having None arguments."""
-        chat_msg = {
-            "role": "assistant",
-            "tool_calls": [
-                {
-                    "function": {
-                        "name": "get_current_time",
-                        "arguments": None,
-                    }
-                }
-            ],
-        }
-
-        messages = parse_function(chat_msg)
-
-        assert len(messages) == 1
-        assert messages[0].content[0].text == ""
-        assert messages[0].recipient == "functions.get_current_time"
-
-    def test_system_message(self, parse_function):
-        """Test parsing system message."""
-        chat_msg = {
-            "role": "system",
-            "content": "You are a helpful assistant",
-        }
-
-        messages = parse_function(chat_msg)
-
-        assert len(messages) == 1
-        # System messages are converted using Message.from_dict
-        # which should preserve the role
-        assert messages[0].author.role == Role.SYSTEM
-
-    def test_developer_message(self, parse_function):
-        """Test parsing developer message."""
-        chat_msg = {
-            "role": "developer",
-            "content": "Use concise language",
-        }
-
-        messages = parse_function(chat_msg)
-
-        assert len(messages) == 1
-        assert messages[0].author.role == Role.DEVELOPER
-
-    def test_user_message_with_string_content(self, parse_function):
-        """Test parsing user message with string content."""
-        chat_msg = {
-            "role": "user",
-            "content": "What's the weather in San Francisco?",
-        }
-
-        messages = parse_function(chat_msg)
-
-        assert len(messages) == 1
-        assert messages[0].author.role == Role.USER
-        assert messages[0].content[0].text == "What's the weather in San Francisco?"
-
-    def test_user_message_with_array_content(self, parse_function):
-        """Test parsing user message with array content."""
-        chat_msg = {
-            "role": "user",
-            "content": [
-                {"text": "What's in this image? "},
-                {"text": "Please describe it."},
-            ],
-        }
-
-        messages = parse_function(chat_msg)
-
-        assert len(messages) == 1
-        assert messages[0].author.role == Role.USER
-        assert len(messages[0].content) == 2
-        assert messages[0].content[0].text == "What's in this image? "
-        assert messages[0].content[1].text == "Please describe it."
-
-    def test_assistant_message_with_string_content(self, parse_function):
-        """Test parsing assistant message with string content (no tool calls)."""
-        chat_msg = {
-            "role": "assistant",
-            "content": "Hello! How can I help you today?",
-        }
-
-        messages = parse_function(chat_msg)
-
-        assert len(messages) == 1
-        assert messages[0].author.role == Role.ASSISTANT
-        assert messages[0].content[0].text == "Hello! How can I help you today?"
-
-    def test_pydantic_model_input(self, parse_function):
-        """Test parsing Pydantic model input (has model_dump method)."""
-
-        class MockPydanticModel:
-            def model_dump(self, exclude_none=True):
-                return {
-                    "role": "user",
-                    "content": "Test message",
-                }
-
-        chat_msg = MockPydanticModel()
-        messages = parse_function(chat_msg)
-
-        assert len(messages) == 1
-        assert messages[0].author.role == Role.USER
-        assert messages[0].content[0].text == "Test message"
-
-    def test_tool_call_with_missing_function_fields(self, parse_function):
-        """Test parsing tool call with missing name or arguments."""
-        chat_msg = {
-            "role": "assistant",
-            "tool_calls": [
-                {
-                    "function": {}  # Missing both name and arguments
-                }
-            ],
-        }
-
-        messages = parse_function(chat_msg)
-
-        assert len(messages) == 1
-        assert messages[0].recipient == "functions."
-        assert messages[0].content[0].text == ""
-
-    def test_array_content_with_missing_text(self, parse_function):
-        """Test parsing array content where text field is missing."""
-        chat_msg = {
-            "role": "user",
-            "content": [
-                {},  # Missing text field
-                {"text": "actual text"},
-            ],
-        }
-
-        messages = parse_function(chat_msg)
-
-        assert len(messages) == 1
-        assert len(messages[0].content) == 2
-        assert messages[0].content[0].text == ""
-        assert messages[0].content[1].text == "actual text"
 
 
 class TestParseChatInputToHarmonyMessage:
     """
     Tests for scenarios that are specific to the Chat Completion API
-    parse_chat_input_to_harmony_message function.
+    chat_input_to_harmony function.
     """
 
     def test_user_message_with_empty_content(self):
@@ -223,7 +26,7 @@ class TestParseChatInputToHarmonyMessage:
             "content": "",
         }
 
-        messages = parse_chat_input_to_harmony_message(chat_msg)
+        messages = chat_input_to_harmony(chat_msg)
 
         verify_harmony_messages(
             messages,
@@ -241,7 +44,7 @@ class TestParseChatInputToHarmonyMessage:
             "content": None,
         }
 
-        messages = parse_chat_input_to_harmony_message(chat_msg)
+        messages = chat_input_to_harmony(chat_msg)
 
         verify_harmony_messages(
             messages,
@@ -259,7 +62,7 @@ class TestParseChatInputToHarmonyMessage:
             "content": "",
         }
 
-        messages = parse_chat_input_to_harmony_message(chat_msg)
+        messages = chat_input_to_harmony(chat_msg)
 
         assert len(messages) == 0
 
@@ -269,7 +72,7 @@ class TestParseChatInputToHarmonyMessage:
             "content": None,
         }
 
-        messages = parse_chat_input_to_harmony_message(chat_msg)
+        messages = chat_input_to_harmony(chat_msg)
 
         assert len(messages) == 0
 
@@ -280,7 +83,7 @@ class TestParseChatInputToHarmonyMessage:
             "reasoning": "",
         }
 
-        messages = parse_chat_input_to_harmony_message(chat_msg)
+        messages = chat_input_to_harmony(chat_msg)
 
         verify_harmony_messages(
             messages,
@@ -300,7 +103,7 @@ class TestParseChatInputToHarmonyMessage:
             "content": "",
         }
 
-        messages = parse_chat_input_to_harmony_message(chat_msg)
+        messages = chat_input_to_harmony(chat_msg)
 
         verify_harmony_messages(
             messages,
@@ -320,7 +123,7 @@ class TestParseChatInputToHarmonyMessage:
             "content": None,
         }
 
-        messages = parse_chat_input_to_harmony_message(chat_msg)
+        messages = chat_input_to_harmony(chat_msg)
 
         verify_harmony_messages(
             messages,
@@ -346,7 +149,7 @@ class TestParseChatInputToHarmonyMessage:
             ],
         }
 
-        messages = parse_chat_input_to_harmony_message(chat_msg)
+        messages = chat_input_to_harmony(chat_msg)
 
         verify_harmony_messages(
             messages,
@@ -375,7 +178,7 @@ class TestParseChatInputToHarmonyMessage:
             "content": "I'll call the tool.",
         }
 
-        messages = parse_chat_input_to_harmony_message(chat_msg)
+        messages = chat_input_to_harmony(chat_msg)
 
         verify_harmony_messages(
             messages,
@@ -409,7 +212,7 @@ class TestParseChatInputToHarmonyMessage:
             "reasoning": "I should use the get_weather tool.",
         }
 
-        messages = parse_chat_input_to_harmony_message(chat_msg)
+        messages = chat_input_to_harmony(chat_msg)
 
         verify_harmony_messages(
             messages,
@@ -444,7 +247,7 @@ class TestParseChatInputToHarmonyMessage:
             "content": "I'll call the tool.",
         }
 
-        messages = parse_chat_input_to_harmony_message(chat_msg)
+        messages = chat_input_to_harmony(chat_msg)
 
         verify_harmony_messages(
             messages,
@@ -479,9 +282,7 @@ class TestParseChatInputToHarmonyMessage:
             "content": "The weather in San Francisco is sunny, 72°F",
         }
 
-        messages = parse_chat_input_to_harmony_message(
-            chat_msg, tool_id_names=tool_id_names
-        )
+        messages = chat_input_to_harmony(chat_msg, tool_id_names=tool_id_names)
 
         verify_harmony_messages(
             messages,
@@ -513,9 +314,7 @@ class TestParseChatInputToHarmonyMessage:
             ],
         }
 
-        messages = parse_chat_input_to_harmony_message(
-            chat_msg, tool_id_names=tool_id_names
-        )
+        messages = chat_input_to_harmony(chat_msg, tool_id_names=tool_id_names)
 
         verify_harmony_messages(
             messages,
@@ -539,9 +338,7 @@ class TestParseChatInputToHarmonyMessage:
             "content": "",
         }
 
-        messages = parse_chat_input_to_harmony_message(
-            chat_msg, tool_id_names=tool_id_names
-        )
+        messages = chat_input_to_harmony(chat_msg, tool_id_names=tool_id_names)
 
         verify_harmony_messages(
             messages,
@@ -565,9 +362,7 @@ class TestParseChatInputToHarmonyMessage:
             "content": None,
         }
 
-        messages = parse_chat_input_to_harmony_message(
-            chat_msg, tool_id_names=tool_id_names
-        )
+        messages = chat_input_to_harmony(chat_msg, tool_id_names=tool_id_names)
 
         verify_harmony_messages(
             messages,
@@ -806,38 +601,3 @@ class TestParseChatOutput:
         reasoning, final_content, _ = parse_chat_output(token_ids)
         assert reasoning is None
         assert final_content == "Let me look that up.\nThe answer is 42."
-
-
-def test_has_custom_tools() -> None:
-    assert not has_custom_tools(set())
-    assert not has_custom_tools({"web_search_preview", "code_interpreter", "container"})
-    assert has_custom_tools({"others"})
-    assert has_custom_tools(
-        {"web_search_preview", "code_interpreter", "container", "others"}
-    )
-
-
-class TestGetSystemMessage:
-    """Tests for get_system_message channel configuration."""
-
-    def test_commentary_channel_present_without_custom_tools(self) -> None:
-        """Commentary channel must be valid even without custom tools."""
-        sys_msg = get_system_message(with_custom_tools=False)
-        valid_channels = sys_msg.content[0].channel_config.valid_channels
-        assert "commentary" in valid_channels
-
-    def test_commentary_channel_present_with_custom_tools(self) -> None:
-        """Commentary channel present when custom tools are enabled."""
-        sys_msg = get_system_message(with_custom_tools=True)
-        valid_channels = sys_msg.content[0].channel_config.valid_channels
-        assert "commentary" in valid_channels
-
-    def test_all_standard_channels_present(self) -> None:
-        """All three standard Harmony channels should always be valid."""
-        for with_tools in (True, False):
-            sys_msg = get_system_message(with_custom_tools=with_tools)
-            valid_channels = sys_msg.content[0].channel_config.valid_channels
-            for channel in ("analysis", "commentary", "final"):
-                assert channel in valid_channels, (
-                    f"{channel} missing when with_custom_tools={with_tools}"
-                )
