+++ vllm/entrypoints/openai/chat_completion/harmony.py
@@ -1,179 +1,31 @@
 # SPDX-License-Identifier: Apache-2.0
 # SPDX-FileCopyrightText: Copyright contributors to the vLLM project
+"""
+Harmony ↔ Chat Completion API conversion utilities.
 
-import datetime
+Handles two directions:
+  1. Chat Completion Input → Harmony Messages  (input parsing)
+  2. Harmony Messages → Chat Completion Output  (output parsing)
+"""
+
 from collections.abc import Iterable, Sequence
-from typing import Literal
 
-from openai.types.responses.tool import Tool
 from openai_harmony import (
     Author,
-    Conversation,
-    DeveloperContent,
-    HarmonyEncodingName,
     Message,
-    ReasoningEffort,
     Role,
     StreamableParser,
-    SystemContent,
     TextContent,
-    ToolDescription,
-    load_harmony_encoding,
 )
 
-from vllm import envs
-from vllm.entrypoints.openai.chat_completion.protocol import ChatCompletionToolsParam
-from vllm.logger import init_logger
+from vllm.entrypoints.openai.harmony import (
+    flatten_chat_text_content,
+    get_streamable_parser_for_assistant,
+)
 
-logger = init_logger(__name__)
-
-REASONING_EFFORT = {
-    "high": ReasoningEffort.HIGH,
-    "medium": ReasoningEffort.MEDIUM,
-    "low": ReasoningEffort.LOW,
-}
-
-_harmony_encoding = None
-
-# Builtin tools that should be included in the system message when
-# they are available and requested by the user.
-# Tool args are provided by MCP tool descriptions. Output
-# of the tools are stringified.
-BUILTIN_TOOL_TO_MCP_SERVER_LABEL: dict[str, str] = {
-    "python": "code_interpreter",
-    "browser": "web_search_preview",
-    "container": "container",
-}
-
-# Derive MCP_BUILTIN_TOOLS from the canonical mapping
-MCP_BUILTIN_TOOLS: set[str] = set(BUILTIN_TOOL_TO_MCP_SERVER_LABEL.values())
-
-
-def has_custom_tools(tool_types: set[str]) -> bool:
-    """
-    Checks if the given tool types are custom tools
-    (i.e. any tool other than MCP buildin tools)
-    """
-    return not tool_types.issubset(MCP_BUILTIN_TOOLS)
-
-
-def get_encoding():
-    global _harmony_encoding
-    if _harmony_encoding is None:
-        _harmony_encoding = load_harmony_encoding(HarmonyEncodingName.HARMONY_GPT_OSS)
-    return _harmony_encoding
-
-
-def get_system_message(
-    model_identity: str | None = None,
-    reasoning_effort: Literal["high", "medium", "low"] | None = None,
-    start_date: str | None = None,
-    browser_description: str | None = None,
-    python_description: str | None = None,
-    container_description: str | None = None,
-    instructions: str | None = None,
-    with_custom_tools: bool = False,
-) -> Message:
-    sys_msg_content = SystemContent.new()
-    if model_identity is not None:
-        sys_msg_content = sys_msg_content.with_model_identity(model_identity)
-    if instructions is not None and envs.VLLM_GPT_OSS_HARMONY_SYSTEM_INSTRUCTIONS:
-        current_identity = sys_msg_content.model_identity
-        new_identity = (
-            f"{current_identity}\n{instructions}" if current_identity else instructions
-        )
-        sys_msg_content = sys_msg_content.with_model_identity(new_identity)
-    if reasoning_effort is not None:
-        sys_msg_content = sys_msg_content.with_reasoning_effort(
-            REASONING_EFFORT[reasoning_effort]
-        )
-    if start_date is None:
-        # NOTE(woosuk): This brings non-determinism in vLLM.
-        # Set VLLM_SYSTEM_START_DATE to pin it.
-        start_date = envs.VLLM_SYSTEM_START_DATE or datetime.datetime.now().strftime(
-            "%Y-%m-%d"
-        )
-    sys_msg_content = sys_msg_content.with_conversation_start_date(start_date)
-    if browser_description is not None:
-        sys_msg_content = sys_msg_content.with_tools(browser_description)
-    if python_description is not None:
-        sys_msg_content = sys_msg_content.with_tools(python_description)
-    if container_description is not None:
-        sys_msg_content = sys_msg_content.with_tools(container_description)
-    sys_msg = Message.from_role_and_content(Role.SYSTEM, sys_msg_content)
-    return sys_msg
-
-
-def create_tool_definition(tool: ChatCompletionToolsParam | Tool):
-    if isinstance(tool, ChatCompletionToolsParam):
-        return ToolDescription.new(
-            name=tool.function.name,
-            description=tool.function.description,
-            parameters=tool.function.parameters,
-        )
-    return ToolDescription.new(
-        name=tool.name,
-        description=tool.description,
-        parameters=tool.parameters,
-    )
-
-
-def get_developer_message(
-    instructions: str | None = None,
-    tools: list[Tool | ChatCompletionToolsParam] | None = None,
-) -> Message:
-    dev_msg_content = DeveloperContent.new()
-    if instructions is not None and not envs.VLLM_GPT_OSS_HARMONY_SYSTEM_INSTRUCTIONS:
-        dev_msg_content = dev_msg_content.with_instructions(instructions)
-    if tools is not None:
-        function_tools: list[Tool | ChatCompletionToolsParam] = []
-        for tool in tools:
-            if tool.type in (
-                "web_search_preview",
-                "code_interpreter",
-                "container",
-            ):
-                pass
-
-            elif tool.type == "function":
-                function_tools.append(tool)
-            else:
-                raise ValueError(f"tool type {tool.type} not supported")
-        if function_tools:
-            function_tool_descriptions = [
-                create_tool_definition(tool) for tool in function_tools
-            ]
-            dev_msg_content = dev_msg_content.with_function_tools(
-                function_tool_descriptions
-            )
-    dev_msg = Message.from_role_and_content(Role.DEVELOPER, dev_msg_content)
-    return dev_msg
-
-
-def get_user_message(content: str) -> Message:
-    return Message.from_role_and_content(Role.USER, content)
-
-
-def parse_chat_inputs_to_harmony_messages(chat_msgs: list) -> list[Message]:
-    """
-    Parse a list of messages from request.messages in the Chat Completion API to
-    Harmony messages.
-    """
-    msgs: list[Message] = []
-    tool_id_names: dict[str, str] = {}
-
-    # Collect tool id to name mappings for tool response recipient values
-    for chat_msg in chat_msgs:
-        for tool_call in chat_msg.get("tool_calls", []):
-            tool_id_names[tool_call.get("id")] = tool_call.get("function", {}).get(
-                "name"
-            )
-
-    for chat_msg in chat_msgs:
-        msgs.extend(parse_chat_input_to_harmony_message(chat_msg, tool_id_names))
-
-    msgs = auto_drop_analysis_messages(msgs)
-    return msgs
+# ---------------------------------------------------------------------------
+# Input parsing: Chat Completion messages → Harmony Messages
+# ---------------------------------------------------------------------------
 
 
 def auto_drop_analysis_messages(msgs: list[Message]) -> list[Message]:
@@ -208,21 +60,7 @@ def auto_drop_analysis_messages(msgs: list[Message]) -> list[Message]:
     return cleaned_msgs
 
 
-def flatten_chat_text_content(content: str | list | None) -> str | None:
-    """
-    Extract the text parts from a chat message content field and flatten them
-    into a single string.
-    """
-    if isinstance(content, list):
-        return "".join(
-            item.get("text", "")
-            for item in content
-            if isinstance(item, dict) and item.get("type") == "text"
-        )
-    return content
-
-
-def parse_chat_input_to_harmony_message(
+def chat_input_to_harmony(
     chat_msg, tool_id_names: dict[str, str] | None = None
 ) -> list[Message]:
     """
@@ -293,8 +131,6 @@ def parse_chat_input_to_harmony_message(
 
     # Default: user/assistant/system messages with content
     content = chat_msg.get("content") or ""
-    if content is None:
-        content = ""
     if isinstance(content, str):
         contents = [TextContent(text=content)]
     else:
@@ -316,20 +152,31 @@ def parse_chat_input_to_harmony_message(
     return msgs
 
 
-def render_for_completion(messages: list[Message]) -> list[int]:
-    conversation = Conversation.from_messages(messages)
-    token_ids = get_encoding().render_conversation_for_completion(
-        conversation, Role.ASSISTANT
-    )
-    return token_ids
+def chat_inputs_to_harmony(chat_msgs: list) -> list[Message]:
+    """
+    Parse a list of messages from request.messages in the Chat Completion API to
+    Harmony messages.
+    """
+    msgs: list[Message] = []
+    tool_id_names: dict[str, str] = {}
+
+    # Collect tool id to name mappings for tool response recipient values
+    for chat_msg in chat_msgs:
+        for tool_call in chat_msg.get("tool_calls", []):
+            tool_id_names[tool_call.get("id")] = tool_call.get("function", {}).get(
+                "name"
+            )
+
+    for chat_msg in chat_msgs:
+        msgs.extend(chat_input_to_harmony(chat_msg, tool_id_names))
+
+    msgs = auto_drop_analysis_messages(msgs)
+    return msgs
 
 
-def get_stop_tokens_for_assistant_actions() -> list[int]:
-    return get_encoding().stop_tokens_for_assistant_actions()
-
-
-def get_streamable_parser_for_assistant() -> StreamableParser:
-    return StreamableParser(get_encoding(), role=Role.ASSISTANT)
+# ---------------------------------------------------------------------------
+# Output parsing: Harmony Messages → Chat Completion Output
+# ---------------------------------------------------------------------------
 
 
 def parse_output_into_messages(token_ids: Iterable[int]) -> StreamableParser:
