+++ vllm/tool_parsers/openai_tool_parser.py
@@ -4,6 +4,7 @@ import json
 from collections.abc import Sequence
 from typing import TYPE_CHECKING
 
+from vllm.entrypoints.openai.chat_completion.harmony import parse_output_into_messages
 from vllm.entrypoints.openai.chat_completion.protocol import (
     ChatCompletionRequest,
 )
@@ -13,7 +14,6 @@ from vllm.entrypoints.openai.engine.protocol import (
     FunctionCall,
     ToolCall,
 )
-from vllm.entrypoints.openai.parser.harmony_utils import parse_output_into_messages
 from vllm.logger import init_logger
 from vllm.tool_parsers.abstract_tool_parser import (
     ToolParser,
